fix(core): make ProtocolRegistry thread-safe; zero compile warnings
ProtocolRegistry accessed its unordered_map without any locking, so
concurrent register_handler calls could drop entries (UB) - the
ConcurrentPluginRegistration test intermittently saw 9 handlers
instead of 10. Protect all access with std::shared_mutex (shared for
reads, exclusive for writes) plus a find_handler_unlocked helper so
the recursive lookups inside get_handler_for_url run under one held
lock without re-entrance. protocol_name() is now called outside the
lock. Document the thread-safety contract in the header.

Also clear all -Wall/-Wextra/-Wconversion/-Wsign-conversion warnings
(87 -> 0, verified with and without the BitTorrent plugin):

- protocol_registry/commands: platform-wrap memmem_alt under _WIN32,
  explicit ptrdiff_t/size_t conversions
- incremental_download: explicit streamoff/streamsize casts; void-cast
  pending-implementation parameters
- storage browsers: [[maybe_unused]] params, OpenSSL int casts,
  CURLINFO_CONTENT_LENGTH_DOWNLOAD -> _T variant, replace local
  variadic LOG_* macros with FALCON_LOG_*
- bittorrent: int64_t bencode values cast explicitly; size_t indexing
- tests: drop unused variables, consume [[nodiscard]] results, explicit
  narrowing casts, fix port 99999 overflow in socket_pool_test, and
  wait asynchronously for on_completed in EventListener (events are
  delivered by the dispatcher thread)

TODO hygiene: remove stale markers (http_request placeholder notes,
HLS pipeline TODO that is already implemented); keep and sharpen the
genuine ones (DHT async lookup wiring, Kademlia iterative lookup,
spdlog migration) and track them in todo.md.

+++ packages/libfalcon-drives/tests/unit/search_basic_test.cpp
@@ -35,7 +35,7 @@ public:
             result.size = i * 100 * 1024 * 1024; // 100MB increments
             result.seeds = static_cast<int>((query.limit - i) * 2);
             result.peers = static_cast<int>((query.limit - i) * 3);
-            result.confidence = 0.5 + (i * 0.05);
+            result.confidence = 0.5 + (static_cast<double>(i) * 0.05);
 
             // 应用过滤条件
             if (query.min_size > 0 && result.size < query.min_size) continue;
